py-pyobjc: unbreak building on older systems
Now that we're actually using the right headers on these systems, we
run into the fact that PyObjC uses a deprecated symbol. As the build
system passes -Werror when compiling C sources, that triggers an
error.

The breakage affects users, so I've applied the quick fix to simply
suppress the error and live with the warning.

Fixes: https://trac.macports.org/ticket/63084